splitting graph_definition
